Initial works on translation
I opted to John Resig's i18n-node-2 because it's very light weight. It's very simplistic when it comes to pluralisation and gender and it may cause trouble, but I think I can overcome this shortcomings.

ATM `pl` locale is hardcoded into `views/components/translate.coffee`. It should be configurable, but that is very low priority right now.